Simplified PKGBUILDs and cleaned up Makefile

+++ .drone/scripts/user-repo.sh
@@ -1,8 +1,26 @@
 #!/usr/bin/env bash
-set -exuo pipefail
+set -e
 sudo chown 'makedeb:makedeb' ./ -R
 
-# Set up SSH
+# Get package version.
+read PKGVER PKGREL < <(echo "${DRONE_COMMIT_MESSAGE}" | grep '^Package Version:' | awk -F ': ' '{print $2}' | sed 's|-| |g')
+
+for i in PKGVER PKGREL; do
+	if [[ "${!i}" == "" ]]; then
+		echo "ERROR: ${i} isn't set."
+		echo "Please make sure your commit message contained a 'Package Version' line."
+		bad_commit_message="x"
+	fi
+done
+
+if [[ "${bad_commit_meesage:+x}" == "x" ]]; then
+	echo "COMMIT MESSAGE:"
+	echo "==============="
+	echo "${DRONE_COMMIT_MESSAGE}"
+	exit 1
+fi
+
+# Set up SSH.
 rm -rf "/${HOME}/.ssh"
 mkdir -p "/${HOME}/.ssh"
 
@@ -23,28 +41,25 @@ fi
 
 chmod 500 "/${HOME}/.ssh/"* -R
 
-# Clone AUR/MPR Package
+# Clone AUR/MPR Package.
 if [[ "${target_repo}" == "mpr" ]]; then
 	git clone "ssh://mpr@${mpr_url}/${package_name}.git" "${package_name}_${target_repo}"
-
 else
 	git clone "ssh://aur@${aur_url}/${package_name}.git" "${package_name}_${target_repo}"
-
 fi
 
-# Copy PKGBUILD to user repo
+# Copy PKGBUILD to user repo.
+export TARGET="${target_repo}"
+export RELEASE="${release_type}"
+export PKGVER
+export PKGREL
+
 rm "${package_name}_${target_repo}/PKGBUILD"
-cp "PKGBUILDs/${target_repo^^}/${release_type^^}.PKGBUILD" "${package_name}_${target_repo}/PKGBUILD"
-
-# Get current pkgver and pkgrel
-pkgver="$(cat src/PKGBUILD | grep '^pkgver=' | awk -F '=' '{print $2}')"
-pkgrel="$(cat src/PKGBUILD | grep '^pkgrel=' | awk -F '=' '{print $2}')"
-
-# Set package version in PKGBUILD
-sed -i "s|^pkgver={pkgver}|pkgver=${pkgver}|" "${package_name}_${target_repo}/PKGBUILD"
+cd PKGBUILD/
+./pkgbuild.sh > "../${package_name}_${target_repo}/PKGBUILD"
 
 # Create .SRCINFO file
-cd "${package_name}_${target_repo}"
+cd "../${package_name}_${target_repo}"
 
 makedeb --printsrcinfo | tee .SRCINFO
 
@@ -60,6 +75,6 @@ git config user.email "kavplex@hunterwittenborn.com"
 
 # Commit changes and push
 git add PKGBUILD .SRCINFO
-git commit -m "Updated version to ${pkgver}-${pkgrel}"
+git commit -m "Updated version to ${PKGVER}-${PKGREL}"
 
 git push
